requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
filelock==3.18.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
narwhals==1.35.0
numpy==2.2.4
openai==1.75.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.14.0
sortedcontainers==2.4.0
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
